marine_base/requirements.txt
streamlit
ultralytics
opencv-python-headless
supabase
python-dotenv
numpy
pandas
langchain
langchain-openai
langchain-anthropic
tiktoken
reportlab
